[Refactor] 동아리 가입신청 페이지 컴포넌트 구조 개선
