+++ src/app/(route)/join/_data/formOptions.ts
@@ -1,0 +1,29 @@
+/**
+ * @module formOptions
+ * @description 가입 폼에서 사용하는 셀렉트 옵션 상수 모음
+ */
+
+/** 성별 옵션 */
+export const GENDER_OPTIONS = [
+  { value: "", label: "성별 선택" },
+  { value: "male", label: "남자" },
+  { value: "female", label: "여자" },
+];
+
+/** 학년 옵션 */
+export const GRADE_OPTIONS = [
+  { value: "", label: "학년 선택" },
+  { value: "1", label: "1학년" },
+  { value: "2", label: "2학년" },
+  { value: "3", label: "3학년" },
+  { value: "4", label: "4학년" },
+  { value: "5", label: "5학년" },
+  { value: "6", label: "6학년" },
+];
+
+/** 학기 옵션 */
+export const SEMESTER_OPTIONS = [
+  { value: "", label: "학기 선택" },
+  { value: "1", label: "1학기" },
+  { value: "2", label: "2학기" },
+];
